requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
streamlit
openai
faiss-cpu